01.c float

+++ 07.Tasks/01.c
@@ -11,9 +11,9 @@ void scan_arr(int arr[], int size)
 }
 
 
-int mean_arr(int arr[], int size)
+float mean_arr(int arr[], int size)
 {
-    int sum = 0;
+    float sum = 0;
     for (int i = 0; i < size; i++)
     {
         sum += arr[i];
@@ -26,7 +26,7 @@ int main()
     int arr[SIZE] = {NULL};
     printf("Enter %d numbers with spaces: ", SIZE);
     scan_arr(arr, SIZE);
-    printf("%.3f", (float)mean_arr(arr, 5));
+    printf("%.3f", mean_arr(arr, 5));
     printf("\n");
     return 0;
 }
